Switch get_score to u64
reason for now is that Wasm can't handle u128

+++ src/game.rs
@@ -16,7 +16,7 @@ impl Game {
         Self { board }
     }
 
-    pub fn get_score(&self) -> u128 {
+    pub fn get_score(&self) -> u64 {
         0
     }
